% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{abun_summary}
\alias{abun_summary}
\title{Summarize abundance}
\usage{
abun_summary(x)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}
}
\description{
Summarize abundance
}
\keyword{utilties}
