% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{add_refseq}
\alias{add_refseq}
\title{Add refseq slot for dada2 based phyloseq object}
\usage{
add_refseq(x, tag = "ASV")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object with seqs as rownames.}

\item{tag}{Provide name for Ids, Default="ASV".}
}
\value{
\code{\link{phyloseq-class}} object
}
\description{
Utility to add refseq slot for dada2 based phyloseq object. Here, the
rownames which are unique sequences, are stored in refseq slot of phyloseq. Sequence 
ids are converted to ids using tag option.
}
\examples{

# ps <- add_refseq(p0,tag="ASV")
# ps

}
\keyword{utilities}
