% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biogeogut}
\alias{biogeogut}
\title{Comparison of small intestine and stool microbiota}
\format{An object of class \code{"phyloseq"}.}
\usage{
data("biogeogut")
}
\description{
Data from a Shetty SA, et al.
Bacterial community was profilled using V4 (EMP primers) and analysed with QIIME. Closed ref based OTU picking.
Taxonomic assignment was done using RDP classifier (80% threshold, Wang method).
}
\examples{
\dontrun{
library(microbiomeutilities)
data("biogeogut")
pseq <- biogeogut
print(biogeogut)
          }
          
}
\keyword{datasets}
