% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominant_taxa.R
\name{dominant_taxa}
\alias{dominant_taxa}
\title{Dominant Taxa}
\usage{
dominant_taxa(x, level = NULL, group = NULL)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{level}{Taxonomic level uses microbiome::aggregate_taxa}

\item{group}{Provide overview by groups. Default=NULL}
}
\value{
A list of two data frames/tibbles
}
\description{
Identify dominant taxa in each sample and give overview.
}
\details{
Identifies the dominant taxa in each sample and gives an overview of frequency
and percent sample that are dominated by each taxon. Can be group wise or overall.
}
\examples{
library(microbiomeutilities)
library(dplyr)
data("zackular2014")
p0 <- zackular2014
x.d <- dominant_taxa(p0, level = "Genus", group = "DiseaseState")
head(x.d$dominant_overview)
}
\keyword{utilities}
