% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_to_besthit.R
\name{format_to_besthit}
\alias{format_to_besthit}
\title{Formatting the \code{\link{phyloseq-class}} Object advanced}
\usage{
format_to_besthit(x, prefix = NULL)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{prefix}{Prefered prefix e.g. OTU-d__denovo161:Roseburia or ASV-d__denovo161:Roseburia}
}
\value{
\code{\link{phyloseq-class}} object
}
\description{
Format the phyloseq object to add the best taxonomy in phyloseq object (tax_table and otu_table).
}
\details{
Most commonly it is observed that the taxonomy file has classification until a given taxonomic level.
         row.names for both tax_table and otu_table have best hit, until maximun genus level (species classification with short amplicons is a myth)is made available. This code is a
         slight modification of the code from  \pkg{ampvis} \code{\link{phyloseq-class}}.
         Here, we directly take the phyloseq object as input and make the necessary formatting.
}
\examples{
\dontrun{
# Example data
library(microbiome)
library(microbiomeutilities)
library(dplyr)
data("zackular2014")
p0 <- zackular2014
p0.f <- format_to_besthit(p0, prefix="OTU-")
}
}
\keyword{utilities}
