% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_abundances.R
\name{get_group_abundances}
\alias{get_group_abundances}
\title{Taxa abundance summary by group}
\usage{
get_group_abundances(x, level, group, transform = "compositional")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{level}{Taxonomic level uses microbiome::aggregate_taxa, if NULL with return OTU/ASVs
level stats.}

\item{group}{Provide overview by groups. Default=NULL and will return values for entire
dataset, akin to taxa_summary.}

\item{transform}{Either "compositional" or "counts". Default= compositional}
}
\value{
A data frames/ grouped tibble
}
\description{
Taxa abundance summary by group. Useful for description of microbiome.
}
\examples{
\dontrun{
library(microbiomeutilities)
data("zackular2014")
p0 <- zackular2014
get_group_abundances(p0, level = "Phylum", group = "DiseaseState")
}
}
\keyword{utilities}
