% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{get_tibble}
\alias{get_tibble}
\title{Convert Phyloseq Slots to Tibbles}
\usage{
get_tibble(x, slot = "otu_table", column_id = "column_id")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{slot}{Must be one of c("otu_table", "sam_data", "tax_table").
Default= "otu_table"}

\item{column_id}{Provide name for the column which will hold the rownames of slot.}
}
\value{
A tibble
}
\description{
Utility to convert phyloseq slots to tibbles.
}
\examples{
library(microbiomeutilities)
data("zackular2014")
p0 <- zackular2014
otu_tibble <- get_tibble(p0,slot="otu_table",column_id="OTUID")
head(otu_tibble)
}
\keyword{utilities}
