% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmp2}
\alias{hmp2}
\title{Test data 2}
\format{
An object of class \code{"phyloseq"}.
}
\usage{
data("hmp2")
}
\description{
Data from a Stansfield J, Dozmorov M. "16s rRNA sequencing data 
from the Human Microbiome Project 2". Randomly choosen 13 particpants
with multiple timepoints for rectum samples.
}
\examples{
\dontrun{
library(microbiomeutilities)
data("hmp2")
pseq <- hmp2
print(hmp2)
}

}
\references{
\itemize{
\item{}{Stansfield J, Dozmorov M (2019). HMP2Data: 16s rRNA sequencing data 
from the Human Microbiome Project 2. R package version 1.1.0,
\url{	https://bioconductor.org/packages/HMP2Data/}}
}
}
\keyword{datasets}
