% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{join_otu_tax}
\alias{join_otu_tax}
\title{Join otu_table and tax_table to Tibble}
\usage{
join_otu_tax(x, column_id = "OTUID")
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{column_id}{Provide name for the column which will hold the rownames of slot.}
}
\value{
A tibble
}
\description{
Utility to join otu_table and tax_table to tibble.
}
\examples{
library(microbiomeutilities)
data("zackular2014")
p0 <- zackular2014
otu_tax <- join_otu_tax(p0,column_id = "OTUID")
head(otu_tax)
}
\keyword{utilities}
