% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_heatmap.R
\name{microbiome_heatmap}
\alias{microbiome_heatmap}
\title{Heatmap using phyloseq and pheatmap}
\usage{
microbiome_heatmap(phyobj, subset_otu, transformation, VariableA, heatcolors,
  ...)
}
\arguments{
\item{phyobj}{\code{\link{phyloseq-class}} object.}

\item{subset_otu}{either NA or number of Top OTUs to use for plotting.}

\item{transformation}{either 'log10', 'clr','Z', 'compositional', or NA}

\item{VariableA}{main variable of Interest.}

\item{heatcolors}{is the option for colors in \code{\link{pheatmap}}. Default is to use viridis
inferno}

\item{...}{Arguments to be passed (for \code{\link{pheatmap}} pheatmap).}
}
\value{
A \code{\link{pheatmap}} plot object
}
\description{
Plot heatmap using \code{\link{phyloseq-class}} object as input.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    data(DynamicIBD)
    ps0 <- DynamicIBD
    colnames(tax_table(ps)) <- c("Domain", "Phylum", "Order", "Class", "Family", "Rank6" "Rank7")

    p <- microbiome_heatmap(ps, subset_otu = 10,
    VariableA = "ibd_subtype",
    heatcolors = NA)
          }
}
\author{
Sudarshan A. Shetty (sudarshanshetty9@gmail.com)
}
\keyword{utilities}
