% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_pipeline.R
\name{microbiome_pipeline}
\alias{microbiome_pipeline}
\title{Microbiome analysis pipeline}
\usage{
microbiome_pipeline(otufile, mapping, taxonomy, treefilename, type, work_dir,
  out_dir, VariableA, VariableB, UnConstOrd, heatmap, filterCount, filterPrev,
  col.palette, filterpseq, samsize, projectname, author)
}
\arguments{
\item{otufile}{biom object, otu_table in csv format or mothur shared files.}

\item{mapping}{Metadata variable to check for groups based sequencing effort csv format.}

\item{taxonomy}{NULL or csv fomatted file.}

\item{treefilename}{For phylogenetic based diversity analysis "*.tre" file}

\item{type}{"biom", "mothur", "simple" simple is for *.csv file.}

\item{work_dir}{Working directory where input files are stored.}

\item{out_dir}{Output directory where all outputs are to be stored.}

\item{VariableA}{Main variable of interest.}

\item{VariableB}{Secondary variable of interest.}

\item{UnConstOrd}{If unconstrained ordination to be plotted TRUE or FALSE.}

\item{heatmap}{Heatmap or not option is TRUE or FALSE.}

\item{filterCount}{Filter OTUs below this count number.}

\item{filterPrev}{Filter OTUs not detected in less than this percent of samples.}

\item{col.palette}{'Spectral', 'Set2', 'Paired' use any of the RColorbrewer palette depending on number of groups
you have in VaraibleA.}

\item{filterpseq}{TRUE or FALSE.}

\item{samsize}{Number of reads to rarefy your data and save the rarefied phyobject.}

\item{projectname}{Name of the project.}

\item{author}{Name of the author/investigator.}
}
\value{
A HTML report with graphs and data stats which can serve as guide for actual analysis.
}
\description{
Function microbiome_pipeline generates an HTML report with preliminary QC, Alpha Diversity, Ordination and
             Composition analysis of OTU tables. This function save all intermediate files incuding figures and phyloseq
             objects in user specified directory.
}
\examples{
\dontrun{

library(microbiomeutilities)
library(microbiome)
microbiome_pipeline(otufile = "my.biom",
                   mapping = "mymap.csv",
                   taxonomy = NULL,
                   treefilename = "myTree.tre",
                   type = "biom",
                   work_dir = "F:/path/my/input/filefolder",
                   out_dir = "F:/path/to/save/my/files/folder",
                   VariableA = "MC_type1",
                   VariableB = "Region",
                   UnConstOrd = TRUE,
                   heatmap = TRUE,
                   filterCount = 4,
                   filterPrev = 0.01,
                   col.palette = "Paired",
                   filterpseq = TRUE,
                   samsize = NA,
                   projectname = "Mock",
                   author = "Sudarshan")
                   }
}
\author{
Contact: Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{utilities}
