% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_methods.R
\name{peak-methods}
\alias{peak-methods}
\alias{peak_abundance}
\alias{peak_taxonomy}
\alias{peak_sample}
\alias{peak_base}
\alias{peak_abundance,phyloseq-method}
\alias{peak_taxonomy,phyloseq-method}
\alias{peak_sample,phyloseq-method}
\alias{peak_base,ANY-method}
\title{Peak into \code{\link[phyloseq]{phyloseq}} objects}
\usage{
peak_abundance(x, nrows = 1:5, ncols = 1:5)

peak_taxonomy(x, nrows = 1:5, ncols = 1:5)

peak_sample(x, nrows = 1:5, ncols = 1:5)

peak_base(x, nrows = 1:5, ncols = 1:5)

\S4method{peak_abundance}{phyloseq}(x, nrows = 1:5, ncols = 1:5)

\S4method{peak_taxonomy}{phyloseq}(x, nrows = 1:5, ncols = 1:5)

\S4method{peak_sample}{phyloseq}(x, nrows = 1:5, ncols = 1:5)

\S4method{peak_base}{ANY}(x, nrows = 1:5, ncols = 1:5)
}
\arguments{
\item{x}{a
\code{\link[phyloseq]{phyloseq}} or \code{data.frame} or
\code{matrix} object}

\item{nrows}{number of rows, to be specified as numeric e.g. 1, or sequence 
of numeric specified as 1:5. to return first to fifth row.}

\item{ncols}{number of cols, to be specified as numeric e.g. 1, or sequence 
of numeric specified as 1:5 to return first to fifth col.}
}
\value{
Print user specified rows and columns
}
\description{
These functions work on \code{otu_table}, \code{tax_table}, \code{sample_data}
or on \code{data.frame} and \code{matrix}.

\code{peak_abundance} returns, user specified number of rows and columns
for \code{otu_table}.

#' \code{peak_taxonomy} returns, user specified number of rows and columns
for \code{tax_table}.

\code{peak_sample} returns, user specified number of rows and columns
for \code{sample_data}.  

\code{peak_base} returns, user specified number of rows and columns
for \code{data.frame} and \code{matrix}.
}
\examples{
data("zackular2014")

peak_abundance(zackular2014, nrows=1:3, ncols = 1:3)

peak_taxonomy(zackular2014, nrows=1:3, ncols = 1:3)

peak_sample(zackular2014, nrows=1:3, ncols = 1:3)

dat.frm <- meta(zackular2014)
# specify specific columns
peak_base(dat.frm, nrows=1:3, ncols = c(1, 3, 4))

matrix_ab <- abundances(zackular2014)
peak_base(matrix_ab, nrows=1:3, ncols = 1:3)

}
