% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_to_ldf.R
\name{phy_to_ldf}
\alias{phy_to_ldf}
\title{Convert phyloseq object to long data format}
\usage{
phy_to_ldf(x, transform.counts)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{transform.counts}{Data transform to be used in plotting (but not in sample/taxon ordering). The options are 'Z-OTU', 'Z-Sample', 'log10' and 'compositional'. See the \code{\link{transform}} function.}
}
\value{
A data frame in long format with appropriate transfomation if requested.
}
\description{
Faster alternative to psmelt function from \code{\link{phyloseq-class}} object. It is important that the sample name in metadata should have "SampleID" and this should match the sample names in OTU table.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    data("dietswap")
    pseq <- subset_samples(dietswap, group == "DI" & nationality == "AFR")
    pseq_df <- phy_to_ldf(pseq, transform.counts = NULL)
    
          }
}
\keyword{utilities}
