% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abund_prev.R
\name{plot_abund_prev}
\alias{plot_abund_prev}
\title{Mean Abundance-Prevalence relation}
\usage{
plot_abund_prev(
  x,
  lower.conf = 0.025,
  upper.conf = 0.975,
  bs.iter = 99,
  color = "Phylum"
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{lower.conf}{Lower confidence interval =0.025}

\item{upper.conf}{Upper confidence interval =0.975}

\item{bs.iter}{Number of bootstrap iterations =99}

\item{color}{taxa level to color. preferablly at phylum}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plots Mean Abundance-Prevalence for taxa. Mean abundance, mean prevalence,
and upper and lower confidence interval for each taxa is calculated by random subsampling.
}
\details{
Check if there are spurious OTUs/ASVs.
}
\examples{
\dontrun{
# Example data
 library(microbiomeutilities)
 asv_ps <- zackular2014
 asv_ps <- microbiome::transform(asv_ps, "compositional")
 asv_ps <- core(asv_ps,detection = 0.0001, prevalence = 0.5)
 asv_ps <- format_to_besthit(asv_ps)
 set.seed(2349)
 p_v <- plot_abund_prev(asv_ps) + 
      geom_vline(xintercept = 0.75, lty="dashed", alpha=0.7) + 
      geom_hline(yintercept = 0.01,lty="dashed", alpha=0.7) +
      scale_color_brewer(palette = "Paired")
 p_v
}

}
\keyword{utilities}
