% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha_diversities.R
\name{plot_alpha_diversities}
\alias{plot_alpha_diversities}
\title{Creat a plot for alpha diversities calculated using the \code{\link{microbiome}} package}
\usage{
plot_alpha_diversities(
  x,
  type,
  index.val = "all",
  plot.type,
  variableA,
  palette
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{type}{Either alpha (Diversity Index) or dominance    (Dominance Index) or evenness    (Evenness Index)}

\item{index.val}{see global function in \code{\link{microbiome}} package}

\item{plot.type}{Three optons c("stripchart", "boxplot", "violin")}

\item{variableA}{Variable of interested to be checked. This will also be used to color the plot}

\item{palette}{Any of the \code{\link{RColorBrewer}} plettes}
}
\value{
\code{\link{ggplot}} object. This can be further modified using ggpubr
}
\description{
Utility plot function for diversity measures calcualted by \code{\link{microbiome}} package.
}
\details{
Uses the \code{\link{microbiome}} package global function to calculate diversities and then returns
a plot.
}
\examples{
library(microbiome)
library(microbiomeutilities)
data("zackular2014")
p0 <- zackular2014
p <- plot_alpha_diversities(p0,
  type = "dominance",
  index.val = "all",
  plot.type = "stripchart",
  variableA = "DiseaseState",
  palette = "jco"
)

print(p)
}
\keyword{<-}
\keyword{analysis}
\keyword{c("shannon,}
\keyword{index.val}
\keyword{simpson")}
\keyword{visualization}
