% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_area.R
\name{plot_area}
\alias{plot_area}
\title{Area plot}
\usage{
plot_area(
  x,
  xvar = NULL,
  level = NULL,
  facet.by = NULL,
  fill.colors = brewer.pal(6, "Paired"),
  abund.thres = 0.001,
  prev.thres = 0.5,
  ncol = 5,
  nrow = 5
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{xvar}{Column name to plot on x-axis.}

\item{level}{Taxonomic level. OTU/ASV level not supported.}

\item{facet.by}{Column with variable that has multiple measurements.}

\item{fill.colors}{brewer.pal(6,"Paired"). Specify colors.}

\item{abund.thres}{= 0.001 check \code{\link{microbiome}} package aggregate_rare function.}

\item{prev.thres}{= 0.1 check \code{\link{microbiome}} package aggregate_rare function.}

\item{ncol}{wrap, specify number of columns.}

\item{nrow}{wrap, specify number of rows.}
}
\value{
\code{\link{ggplot}} object.
}
\description{
Create an area plot for longitudinal samples 
with \code{\link{ggplot2}} package.
}
\examples{
\dontrun{
library(microbiomeutilities)
data("hmp2")
ps <- hmp2
ps.rel <- microbiome::transform(ps, "compositional") 
p <- plot_area(ps.rel, xvar="visit_number", 
              level = "Phylum",
              facet.by = "subject_id",
              fill.colors=brewer.pal(6,"Paired"))
}           

}
\keyword{visualization}
