% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diversity_stats.R
\name{plot_diversity_stats}
\alias{plot_diversity_stats}
\title{Diversity plot with stats}
\usage{
plot_diversity_stats(
  x,
  index,
  group = NULL,
  group.colors = c("brown3", "steelblue"),
  dot.opacity = 0.25,
  box.opacity = 0.25,
  violin.opacity = 0.5,
  group.order = NULL,
  ...
)
}
\arguments{
\item{x}{phyloseq object}

\item{index}{diversity index. Calculated using microbiome::alpha}

\item{group}{Grouping variable to compare}

\item{group.colors}{Colors for plotting groups}

\item{dot.opacity}{for ggplot alpha to determine opacity for points}

\item{box.opacity}{for ggplot alpha to determine opacity for box}

\item{violin.opacity}{for ggplot alpha to determine opacity for violin}

\item{group.order}{Default is NULL. a list specifing order of x-axis. 
E.g. c("H","CRC","nonCRC")}

\item{...}{params for ggpubr::stat_compare_means}
}
\description{
Diversity plot with stats
}
\examples{
library(microbiomeutilities)
library(ggpubr)
data("zackular2014")
p0 <- zackular2014
mycols <- c("brown3", "steelblue", "grey50")
p.m <- plot_diversity_stats(p0, group = "DiseaseState", 
index = "diversity_shannon", group.order = c("H", "CRC", "nonCRC"), 
group.colors = mycols)
p.m
}
\keyword{utilities}
