% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diversity_stats_utils.R
\name{plot_diversity_with_stats}
\alias{plot_diversity_with_stats}
\title{Diversity plot with stats}
\usage{
plot_diversity_with_stats(
  x,
  index,
  group = NULL,
  group.colors = c("brown3", "steelblue"),
  dot.opacity = 0.25,
  box.opacity = 0.25,
  violin.opacity = 0.5,
  dot.size = 2,
  group.order = NULL,
  ...
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{index}{diversity index. Calculated using microbiome::alpha}

\item{group}{Grouping variable to compare}

\item{group.colors}{Colors for plotting groups}

\item{dot.opacity}{for ggplot alpha to determine opacity for points}

\item{box.opacity}{for ggplot alpha to determine opacity for box}

\item{violin.opacity}{for ggplot alpha to determine opacity for violin}

\item{dot.size}{size of point
E.g. c("H","CRC","nonCRC")}

\item{group.order}{Default is NULL. a list specifing order of x-axis.}

\item{...}{params for ggpubr::stat_compare_means}
}
\description{
Diversity plot with stats
}
\keyword{utilties}
