% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_listed_taxa.R
\name{plot_listed_taxa}
\alias{plot_listed_taxa}
\title{A boxplot for user specified list of taxa}
\usage{
plot_listed_taxa(
  x,
  select.taxa,
  group,
  group.colors,
  dot.opacity = 0.25,
  box.opacity = 0.25,
  group.order = NULL,
  add.violin = TRUE,
  violin.opacity = 0.25,
  panel.arrange = "grid",
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{select.taxa}{a character list of taxa to be plotted. eg. select.taxa <- c("OTU-370251", "OTU-311173", "OTU-341024").}

\item{group}{Grouping variable to compare}

\item{group.colors}{Colors for plotting groups}

\item{dot.opacity}{For ggplot alpha to determine opacity for points}

\item{box.opacity}{For ggplot alpha to determine opacity for box}

\item{group.order}{Default is NULL. a list specifing order of x-axis.}

\item{add.violin}{Loical. If half violoin to the added. Default=TRUE}

\item{violin.opacity}{If add.violin=TRUE, opacity for violin.}

\item{panel.arrange}{panels "grid" or "wrap" ggplot's facet_XXX}

\item{ncol}{if wrap, specify number of columns.}

\item{nrow}{if wrap, specify number of rows.}
}
\value{
\code{\link{ggplot}} object. This can be further modified using ggpubr.
}
\description{
User specifed OTUs are plotted.
}
\details{
Useful for instances where user is interested only in some OTUs. For example OTUs
         reported to be significantly diferent. This can also be used at higher taxonomic
         levels, output from phyloseq::tax_glom or microbiome::aggregate_taxa.
}
\examples{
\dontrun{
# Example data
library(microbiome)
library(microbiomeutilities)
data("zackular2014")
p0 <- zackular2014
p0.f <- format_to_besthit(p0)
select.taxa <- c("OTU-d__denovo31:Dorea", "OTU-d__denovo24:Blautia")
mycols <- c("brown3", "steelblue", "grey50")
p <- plot_listed_taxa(p0.f, select.taxa, 
group="DiseaseState", 
add.violin=TRUE, 
group.colors=mycols)
print(p)
}
}
\keyword{visualization}
