% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordiplot_core.R
\name{plot_ordiplot_core}
\alias{plot_ordiplot_core}
\title{Plotting core microibota on ordinations}
\usage{
plot_ordiplot_core(x, ordiObject, coreplot = c(TRUE, FALSE), prevalences,
  detections, color.opt, shape)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{ordiObject}{Output of ordinate from package phyloseq. Only NMDS and Bray supported.}

\item{coreplot}{TRUE or FALSE for core heatmap from microbiome package.}

\item{prevalences}{Prevalences as supported by microbiome package.}

\item{detections}{Detections as supported by microbiome package.}

\item{color.opt}{Variable of interest from metadata.}

\item{shape}{Variable of interest from metadata.}
}
\value{
plot
}
\description{
This function will plot the ordaination alsong with highligthing the core microbes
on the species ordination.
}
\details{
Most commonly it is observed that the taxonomy file has classification until a given
         taxonomic level.
         Hence, to avoid loss of OTU information while using the function tax_glom() for merging at a specific taxonomic level.
         we will fill the empty cells with the maximum classification available along with the
         OTU number. This code is a slight modification.
         the code from  \pkg{ampvis} \code{\link{phyloseq-class}}. Here, we directly take the
         phyloseq object as input and make the necessary formatting.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    library(microbiomeUtilities)
    library(RColorBrewer)
    data("biogeogut")
    p0 <- biogeogut
    ps1 <- format_phyloseq(ps0)
    px <- plot_ordiplot_core(ps1, ordiObject = ordi, coreplot = TRUE, 
    prevalences = prev.thres,
    detections = det.thres, color.opt = "SampleType", shape = NULL)
          }
}
\keyword{utilities}
