% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_select_taxa.R
\name{plot_select_taxa}
\alias{plot_select_taxa}
\title{A boxplot for user specified list of taxa}
\usage{
plot_select_taxa(x, select.taxa, variableA, palette, plot.type)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{select.taxa}{a character list of taxa to be plotted. eg. select.taxa <- c("OTU-370251", "OTU-311173", "OTU-341024", "OTU-179814").}

\item{variableA}{Variable of interested to be checked. This will also be used to color the plot.}

\item{palette}{Any of the RColorBrewer plettes.}

\item{plot.type}{Three optons c("stripchart", "boxplot", "violin")}
}
\value{
\code{\link{ggplot}} object. This can be further modified using ggpubr.
}
\description{
User specifed OTUs are plotted.
}
\details{
Useful for instances where user is interested only in some OTUs. For example OTUs
         reported to be significantly diferent.
}
\examples{
    library(microbiome)
    library(microbiomeutilities)
    data("biogeogut")
    p0 <- biogeogut
    p0.f <- format_to_besthit(p0)
    select.taxa <- c("OTU-4431292:Oscillospira", "OTU-661266:Blautia")
    p <- plot_select_taxa(p0.f, select.taxa, "SampleType", "Paired", plot.type = "stripchart")
    print(p)

}
\keyword{utilities}
