% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_boxplot.R
\name{plot_taxa_boxplot}
\alias{plot_taxa_boxplot}
\title{Taxonomic Composition Plot boxplot}
\usage{
plot_taxa_boxplot(
  x,
  taxonomic.level,
  top.otu,
  keep.other = FALSE,
  group,
  title,
  group.colors = NULL,
  group.order = NULL,
  add.violin = TRUE,
  violin.opacity = 0.25,
  box.opacity = 0.25,
  dot.opacity = 0.25,
  dot.size = 2
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{taxonomic.level}{Merge the OTUs (for phyloseq object) into a higher taxonomic level. This has to be one from colnames(tax_table(x)).}

\item{top.otu}{Top number of taxa to plot.}

\item{keep.other}{TRUE or FALSE. Default is FALSE. This will not plot taxa group as Other}

\item{group}{Specify main variable of interest. This should be one of the variables in sample_variables(x).}

\item{title}{title for the plot}

\item{group.colors}{Colors for plotting groups}

\item{group.order}{Default is NULL. a list specifing order of x-axis. E.g. c("H","CRC","nonCRC")}

\item{add.violin}{Loical. If half violoin to the added. Default=TRUE}

\item{violin.opacity}{If add.violin=TRUE, opacity for violin.}

\item{box.opacity}{For ggplot alpha to determine opacity for box}

\item{dot.opacity}{For ggplot alpha to determine opacity for points}

\item{dot.size}{For ggplot alpha to determine size for points}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plot taxon abundance for samples.
}
\examples{
\dontrun{
# Example data
library(microbiomeutilities)
library(RColorBrewer)
data("zackular2014")
ps0 <- zackular2014
mycols <- c("brown3", "steelblue", "grey50")
pn <- plot_taxa_boxplot(ps0,
                        taxonomic.level = "Phylum",
                        top.otu = 6, 
                        group = "DiseaseState",
                        title = "Relative abudance plot",
                        keep.other = FALSE,
                        group.order = c("H","CRC","nonCRC"),
                        group.colors = mycols)
print(pn + theme_biome_utils())
}

}
\keyword{visualization}
