% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_composition.R
\name{plot_taxa_composition}
\alias{plot_taxa_composition}
\title{Taxonomic Composition Plot}
\usage{
plot_taxa_composition(x, sample.sort = NULL, taxonomic.level, transform,
  otu.sort = NULL, palette = brewer.pal(12, "Paired"), x.label = "sample",
  plot.type = "barplot", average_by, verbose = FALSE, mar = c(5, 12, 1,
  1), ...)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{sample.sort}{Order samples. Various criteria are available:
\itemize{
  \item NULL or 'none': No sorting
  \item A single character string: indicate the metadata field to be used for ordering
  \item A character vector: sample IDs indicating the sample ordering.
  \item 'neatmap' Order samples based on the neatmap approach. See \code{\link{neatsort}}. By default, 'NMDS' method with 'bray' distance is used. For other options, arrange the samples manually with the function.
}}

\item{taxonomic.level}{Merge the OTUs (for phyloseq object) into a higher taxonomic level. This has to be one from colnames(tax_table(x)).}

\item{transform}{Data transform to be used in plotting (but not in sample/taxon ordering). The options are 'Z-OTU', 'Z-Sample', 'log10' and 'compositional'. See the \code{\link{transform}} function.}

\item{otu.sort}{Order taxa. Same options as for the sample.sort argument but instead of metadata, taxonomic table is used. Also possible to sort by 'abundance'.}

\item{palette}{The number and palette \code{\link{RColorBrewer}} has to be specified e.g brewer.pal(12, "Paired").}

\item{x.label}{Specify how to label the x axis. This should be one of the variables in sample_variables(x).}

\item{plot.type}{Plot type: 'barplot' or 'lineplot'.}

\item{average_by}{Variable to group.}

\item{verbose}{verbose.}

\item{mar}{Figure margins.}

\item{...}{Arguments to be passed (for \code{\link{neatsort}} function)}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plot taxon abundance for samples. It is a legacy function from \code{\link{microbiome}}.
}
\examples{
\dontrun{
    # Example data
    library(microbiome)
    library(microbiomeutilities)
    data("biogeogut")
    pseq <- biogeogut
    plot_taxa_composition(pseq, taxonomic.level = "Phylum")
          }
}
\keyword{utilities}
