% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_cv.R
\name{plot_taxa_cv}
\alias{plot_taxa_cv}
\title{Coefficient of variations}
\usage{
plot_taxa_cv(x)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plots CV for OTUs/ASVs.
}
\details{
Check if there are spurious OTUs/ASVs.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    library(microbiomeUtilities)
    data("biogeogut")
    p0 <- biogeogut
    ps.rel <- microbiome::transform(ps, "compositional")
    p <- plot_taxa_cv(ps.rel)
    print(p)
          }
}
\keyword{utilities}
