% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_heatmap.R
\name{plot_taxa_heatmap}
\alias{plot_taxa_heatmap}
\title{Heatmap using phyloseq and pheatmap}
\usage{
plot_taxa_heatmap(x, subset.top, transformation, VariableA, heatcolors = NULL,
  ...)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{transformation}{either 'log10', 'clr','Z', 'compositional', or NA}

\item{VariableA}{main variable of Interest.}

\item{heatcolors}{is the option for colors in \code{\link{pheatmap}}. Default is to use viridis
inferno}

\item{...}{Arguments to be passed \code{\link{pheatmap}}.}

\item{subset_otu}{either NA or number of Top OTUs to use for plotting.}
}
\value{
A \code{\link{pheatmap}} plot object.
}
\description{
Plot heatmap using \code{\link{phyloseq-class}} object as input.
}
\examples{
\dontrun{
  # Example data
    library(microbiomeutilities)
    library(viridis)
    data("biogeogut")
    ps0 <- biogeogut

    p <- plot_taxa_heatmap(ps, taxonomic.level = "Genus", subset.top = 10,
    VariableA = "SampleType",
    heatcolors = NA)
          }
}
\author{
Sudarshan A. Shetty (sudarshanshetty9@gmail.com)
}
\keyword{utilities}
