% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ternary.R
\name{plot_ternary}
\alias{plot_ternary}
\title{Ternary plot OTU}
\usage{
plot_ternary(
  x,
  abund.thres = 1e-04,
  prev.thres = 0.1,
  group = NULL,
  level = "lowest",
  level.color = "Phylum",
  dot.size = 3,
  dot.opacity = 0.25,
  level.palette = brewer.pal(12, "Paired")
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{abund.thres}{= 0.0001 check \code{\link{microbiome}} package core function
remove taxa that are dectected at 0.0001 in less than prev.thres of samples}

\item{prev.thres}{= 0.1 check \code{\link{microbiome}} package core function}

\item{group}{Grouping variable to compare, for this plot there has to be three 
groups in the data}

\item{level}{= "Genus" Taxonomic level. If OTU/ASV level specify="lowest"
Does not support phylum level aggregation}

\item{level.color}{= "Phylum" Taxonomic level to color}

\item{dot.size}{for ggplot alpha to determine size for points}

\item{dot.opacity}{for ggplot alpha to determine opacity for points}

\item{level.palette}{color for level.color}
}
\value{
\code{\link{ggplot}} object.
}
\description{
Create a ternary plot \code{\link{ggtern}} package.
}
\details{
Plots the mean relative abundance of taxa in 3 groups being compared.
}
\examples{
library(microbiome)
library(microbiomeutilities)
library(dplyr)
data("zackular2014")
p0 <- zackular2014
plot_ternary(p0, group="DiseaseState", 
abund.thres=0.0001, level= "Genus", prev.thres=0.25)
}
\keyword{visualization}
