% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tern_otu.R
\name{prep_tern_otu}
\alias{prep_tern_otu}
\title{Create table for ternary plot OTU}
\usage{
prep_tern_otu(x, abund.thres = 1e-04, prev.thres = 0.1, group = NULL)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{abund.thres}{= 0.0001 check \code{\link{microbiome}} package core function.
remove taxa that are detected at 0.0001 in less than prev.thres of samples.}

\item{prev.thres}{= 0.1 check \code{\link{microbiome}} package core function.}

\item{group}{Grouping variable to compare, for this plot there has to be three
groups in the data}
}
\value{
tibble object.
}
\description{
Create a table for ternary plot ggtern package.
}
\details{
Plots the mean relative abundance of taxa in 3 groups being compared.
}
\examples{
# library(microbiome)
# library(microbiomeutilities)
# library(dplyr)
# data("zackular2014")
# p0 <- zackular2014
# prep_tern_otu(p0, group="DiseaseState",
# abund.thres=0.0001, prev.thres=0.25)
}
\keyword{utilities}
\keyword{visualization}
