% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ternary.R
\name{prep_ternary}
\alias{prep_ternary}
\title{Create table for Ternary plot}
\usage{
prep_ternary(
  x,
  abund.thres = 1e-04,
  prev.thres = 0.1,
  group = NULL,
  level = "lowest"
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{abund.thres}{= 0.0001 check \code{\link{microbiome}} package core function
remove taxa that are dectected at 0.0001 in less than prev.thres of samples}

\item{prev.thres}{= 0.1 check \code{\link{microbiome}} package core function}

\item{group}{Grouping variable to compare, for this plot there has to be three
groups in the data}

\item{level}{= "Genus" Taxonomic level. If OTU/ASV level specify="lowest"
Does not support phylum level aggregation}
}
\value{
Tibble object.
}
\description{
Create a table for ternary plot ggtern R package.
}
\details{
Plots the mean relative abundance of taxa in 3 groups being compared.
}
\examples{
library(microbiome)
library(microbiomeutilities)
library(dplyr)
data("zackular2014")
p0 <- zackular2014
prep_ternary(p0, group = "DiseaseState", abund.thres = 0.0001, level = "Genus", prev.thres = 0.25)
}
\keyword{visualization}
