% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ps.R
\name{print_ps}
\alias{print_ps}
\title{Overview of \code{\link{phyloseq-class}}}
\usage{
print_ps(x)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}
}
\value{
Prints information about the \code{\link{phyloseq-class}} object.
}
\description{
Prints an overview \code{\link{phyloseq-class}}.
}
\examples{
    library(microbiomeutilities)
    data("zackular2014")
    pseq <- zackular2014
    print_ps(pseq)

}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{utilities}
