% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{rarefy_util}
\alias{rarefy_util}
\title{#@param mat otu table rows as taxa names
#@param names_fun either rownames or colnames
#@param list_names vector of names that should appear in boldface}
\usage{
rarefy_util(ps, sample.size, index)
}
\description{
#@param mat otu table rows as taxa names
#@param names_fun either rownames or colnames
#@param list_names vector of names that should appear in boldface
}
\keyword{internal}
