% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microbiome_pipeline.R
\name{run_microbiome_pipeline}
\alias{run_microbiome_pipeline}
\title{Microbiome analysis pipeline}
\usage{
run_microbiome_pipeline(biom, mapping, taxonomy, out_dir, VariableA, VariableB,
  Distmethod, OrdinatMethod, QC, filterCount, filterPrev)
}
\arguments{
\item{biom}{biom object object}

\item{mapping}{Metadata variable to check for groups based sequencing effort.}

\item{taxonomy}{either density or histogram plot}

\item{out_dir}{otuput directory}

\item{VariableA}{Main variable of interest.}

\item{VariableB}{Secondary variable of interest}

\item{Distmethod}{Distance method 'bray', 'unifrac', 'wunifrac', 'jsd'}

\item{OrdinatMethod}{Ordination method 'NMDS', 'MDS', 'PCoA'}

\item{QC}{Default TRUE QC plots}

\item{filterCount}{Filter OTUs below this count number}

\item{filterPrev}{Filter OTUs not detected in less than this percent of samples}

\item{transformation}{'compositional', 'Z', 'log10', 'hellinger', 'clr' choose to tranform your counts data defaults to FALSE}

\item{col.palette}{"Spectral", "Set2", "Paired" use any of the RColorbrewer palette depending on number of groups you have in VaraibleA}
}
\value{
A \code{\link{ggplot}} plot object
}
\description{
This is a single workflow for prelimanry investigation of microbiome data
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    data(DynamicIBD)
    ps0 <- DynamicIBD
    p <- plot_ReadDistribution(ps0, groups='ibd_subtype', plot.type= 'density')
          }
}
\author{
Contact: SUdarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{utilities}
