% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_taxa_plot.R
\name{select_taxa_plot}
\alias{select_taxa_plot}
\title{Formatting the Phyloseq Object advanced}
\usage{
select_taxa_plot(x, select.taxa, variableA, palette, plot.type)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{select.taxa}{a character list of taxa to be plotted. eg. select.taxa <- c("OTU-370251", "OTU-311173", "OTU-341024", "OTU-179814").}

\item{variableA}{Variable of interested to be checked. This will also be used to color the plot.}

\item{palette}{Any of the RColorBrewer plettes.}

\item{plot.type}{Three optons c("stripchart", "boxplot", "violin")}
}
\value{
\code{\link{ggplot}} object. This can be further modified using ggpubr.
}
\description{
Format the phyloseq object to add the best taxonomy in phyloseq object (tax_table and otu_table).
}
\details{
Most commonly it is observed that the taxonomy file has classification until a given taxonomic level.
         row.names for both tax_table and otu_table have best hit, until maximun genus level (species classification with short amplicons is a myth)is made available. This code is a 
         slight modification of the code from  \pkg{ampvis} \code{\link{phyloseq-class}}. 
         Here, we directly take the phyloseq object as input and make the necessary formatting.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    library(microbiomeUtilities)
    data("biogeogut")
    p0 <- biogeogut
    
    select.taxa <- c("OTU-370251:Endozoicimonaceae", "OTU-311173:Helicobacteraceae")
    
    p0.f <- format_to_besthit(p0)
    p <- select_taxa_plot(ps2.f, select.taxa, "SampleType", "Paired", plot.type = "stripchart")
    
    ggpar(p, yscale = "log10", ylab = "Abundance (log10)")
    
          }
}
\keyword{utilities}
