% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_heatmap.R
\name{simple_heatmap}
\alias{simple_heatmap}
\title{Simple Heatmap}
\usage{
simple_heatmap(
  x,
  group.facet = "DiseaseState",
  group.order = c("H", "CRC", "nonCRC"),
  abund.thres = 0.01,
  prev.thres = 0.1,
  level = "Genus",
  scale.color = "log10",
  na.fill = "white",
  color.fill = NULL,
  taxa.arrange = NULL,
  panel.arrange = NULL,
  remove.other = TRUE,
  ncol = NULL,
  nrow = NULL
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{group.facet}{Variable to make facet/panel the plot.}

\item{group.order}{Default is NULL. a list specifing order of x-axis. 
E.g. c("H","CRC","nonCRC")}

\item{abund.thres}{= 0.01 check \code{\link{microbiome}} package aggregate_rare function.}

\item{prev.thres}{= 0.1 check \code{\link{microbiome}} package aggregate_rare function.}

\item{level}{= "Genus" Taxonomic level. OTU/ASV level not supported. Check plot_taxa_heatmap}

\item{scale.color}{Scale the tiles colors "log10" or "sqrt"}

\item{na.fill}{Color to fill NAs. e.g. "white"}

\item{color.fill}{User specified color vectors.}

\item{taxa.arrange}{Arrange the order of taxa. User can supply a list of vectors.}

\item{panel.arrange}{panels "grid" or "wrap" ggplot's facet_XXX}

\item{remove.other}{Rare clubbed as Other to be removed. Logical TRUE/FALSE.}

\item{ncol}{if wrap, specify number of columns.}

\item{nrow}{if wrap, specify number of rows.}
}
\value{
\code{\link{ggplot}} object.
}
\description{
Create a simple heatmap with \code{\link{ggplot2}} package.
}
\details{
Wrapper converts \code{\link{phyloseq-class}} object to long data frame
and generates a heatmap.
}
\examples{
library(microbiome)
library(microbiomeutilities)
library(dplyr)
data("zackular2014")
p0 <- zackular2014
p0.rel <- transform(p0, "compositional")
p <- simple_heatmap(p0.rel, group.facet = "DiseaseState",
                   group.order = c("H", "CRC", "nonCRC"),
                   abund.thres = 0.01,
                   prev.thres = 0.1,
                   level = "Genus",
                   scale.color = "log10",
                   na.fill = "white",
                   color.fill = NULL,
                   taxa.arrange=NULL,
                   remove.other=TRUE,
                   panel.arrange="wrap",
                   ncol=2,
                   nrow=2)

print(p)
}
\keyword{visualization}
