% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{taxa_distribution}
\alias{taxa_distribution}
\title{Distribution of taxa}
\usage{
taxa_distribution(
  x,
  color.level = "Phylum",
  color.taxa = brewer.pal(12, "Paired")
)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{color.level}{Taxonomic level to color}

\item{color.taxa}{vector of colors specified by user
Default is brewer.pal(12,"Paired")}
}
\value{
ggplot2 object
}
\description{
Plots distribution of taxa.
}
\examples{
library(microbiomeutilities)
data("zackular2014")
pseq <- zackular2014
p <- taxa_distribution(pseq)
p
}
\keyword{utilities}
