% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_summary.R
\name{taxa_summary}
\alias{taxa_summary}
\title{Give taxa summary at specified taxonomic level}
\usage{
taxa_summary(x, level)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{level}{Taxonomic level for which summary is required}
}
\value{
returns a data frame with relative abundance summary.
}
\description{
Data frame with mean, max, median standard deviation of relative abundance.
}
\examples{
\dontrun{
    # Example data
    library(microbiomeutilities)
    data("zackular2014")
    p0 <- zackular2014
    p0.rel <- microbiome::transform(p0, "compositional")
    tx.sum1 <- taxa_summary(p0, "Phylum")

    tx.sum2 <- taxa_summary(p0.rel, "Phylum")
    }

}
\author{
Contact: Sudarshan A. Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{utilities}
