% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zackular2014}
\alias{zackular2014}
\title{Test data}
\format{An object of class \code{"phyloseq"}.}
\usage{
data("zackular2014")
}
\description{
Data from a Zackular, Joseph P., et al. "The gut microbiome modulates colon tumorigenesis." MBio 4.6 (2013): e00692
-13.
}
\examples{
\dontrun{
library(microbiomeutilities)
data("zackular2014")
pseq <- zackular2014
print(zackular2014)
}

}
\keyword{datasets}
