% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microbiome_data.R
\name{get_microbiome_data}
\alias{get_microbiome_data}
\title{Download test microbiome data}
\usage{
get_microbiome_data(disease, study)
}
\arguments{
\item{disease}{Disease of interest as shown in list_microbiome_data()}

\item{study}{Name of the study as shown in list_microbiome_data()}
}
\value{
\code{\link{phyloseq-class}} object.
}
\description{
Test microbiome data in phyloseq format.
}
\details{
You can download few example datasets in phyloseq format from Duvallet et al 2017 https://www.nature.com/articles/s41467-017-01973-8.pdf?origin=ppub. The source file for these data is the microbiomedatarepo https://github.com/microsud/microbiomedatarepo.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    library(microbiomeUtilities)
    list_microbiome_data()
    ps1 <- get_microbiome_data(disease = "CDI", "Schubert2014_CDI")
    print_ps(ps1)
          }
}
\keyword{utilities}
