% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_microbiome_data.R
\name{list_microbiome_data}
\alias{list_microbiome_data}
\title{List of available datasets}
\usage{
list_microbiome_data()
}
\description{
Data are used from Duvallet et al 2017 https://www.nature.com/articles/s41467-017-01973-8.pdf?origin=ppub.
}
\details{
Data for practice, also an example for importing mothur files from Baxtrer et al 2016. The source file for these data is the microbiomedatarepo https://github.com/microsud/microbiomedatarepo.
}
\examples{
    library(microbiomeutilities)

    list_microbiome_data()

}
\keyword{utilities}
