% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha_diversities.R
\name{plot_alpha_diversities}
\alias{plot_alpha_diversities}
\title{plot_alpha_diversities}
\usage{
plot_alpha_diversities(x, type, index.val = "all", plot.type, variableA,
  palette)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{type}{Either diversities (Diversity Index) or dominance    (Dominance Index) or evenness    (Evenness Index).}

\item{index.val}{see global function in \code{\link{microbiome}} package.}

\item{plot.type}{Three optons c("stripchart", "boxplot", "violin")}

\item{variableA}{Variable of interested to be checked. This will also be used to color the plot}

\item{palette}{Any of the RColorBrewer plettes.}
}
\value{
\code{\link{ggplot}} object. This can be further modified using ggpubr.
}
\description{
Utility plot function for diversity measures calcualted by \code{\link{microbiome}} package.
}
\details{
Useful for instances where user is interested only in some OTUs. For example OTUs 
         reported to be significantly diferent.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    library(microbiomeUtilities)
    data("biogeogut")
    p0 <- biogeogut
    p <- plot_alpha_diversities(p0, 
    type = "diversities", 
    index.val = "all", 
    plot.type = "stripchart", 
    variableA = "SampleType", 
    palette = "jco")
    
    print(p)
    
          }
}
\keyword{<-}
\keyword{c("shannon,}
\keyword{index.val}
\keyword{simpson")}
\keyword{utilities}
