% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_distribution.R
\name{plot_read_distribution}
\alias{plot_read_distribution}
\title{Distribution of reads}
\usage{
plot_read_distribution(x, groups, plot.type = c("density", "histogram"))
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{groups}{Metadata variable to check for groups based sequencing effort.}

\item{plot.type}{either density or histogram plot}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plots distribution of reads.
}
\examples{
    library(microbiome)
    data(biogeogut)
    ps0 <- biogeogut
    p <- plot_read_distribution(ps0, groups='SampleType', plot.type= 'density')
    print(p)

}
\author{
Contact: Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{utilities}
