% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_select_taxa.R
\name{plot_select_taxa}
\alias{plot_select_taxa}
\title{plot_select_taxa plot selected OTUs of interest}
\usage{
plot_select_taxa(x, select.taxa, variableA, palette, plot.type)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{select.taxa}{a character list of taxa to be plotted. eg. select.taxa <- c("OTU-370251", "OTU-311173", "OTU-341024", "OTU-179814").}

\item{variableA}{Variable of interested to be checked. This will also be used to color the plot.}

\item{palette}{Any of the RColorBrewer plettes.}

\item{plot.type}{Three optons c("stripchart", "boxplot", "violin")}
}
\value{
\code{\link{ggplot}} object. This can be further modified using ggpubr.
}
\description{
User specifed OTUs are plotted.
}
\details{
Useful for instances where user is interested only in some OTUs. For example OTUs 
         reported to be significantly diferent.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    library(microbiomeUtilities)
    data("biogeogut")
    p0 <- biogeogut
    
    select.taxa <- c("OTU-370251:Endozoicimonaceae", "OTU-311173:Helicobacteraceae")
    
    p0.f <- format_to_besthit(p0)
    p <- select_taxa_plot(ps2.f, select.taxa, "SampleType", "Paired", plot.type = "stripchart")
    
    ggpar(p, yscale = "log10", ylab = "Abundance (log10)")
    
          }
}
\keyword{utilities}
