% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_boxplot.R
\name{plot_taxa_boxplot}
\alias{plot_taxa_boxplot}
\title{Taxonomic Composition Plot boxplot}
\usage{
plot_taxa_boxplot(x, taxonomic.level, top.otu, VariableA, title,
  color = NULL)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object}

\item{taxonomic.level}{Merge the OTUs (for phyloseq object) into a higher taxonomic level. This has to be one from colnames(tax_table(x)).}

\item{top.otu}{Top number of taxa to plot.}

\item{VariableA}{Specify main variable of interest. This should be one of the variables in sample_variables(x).}

\item{title}{title for the plot}

\item{color}{any of the palette supported by ggpubr/RColorBrewer packages or  user specified as c("red", "blue").}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plot taxon abundance for samples.
}
\examples{
\dontrun{
# Example data
library(microbiomeutilities)
library(RColorBrewer)
data("zackular2014")
ps1 <- zackular2014
pn <- plot_taxa_boxplot(ps1,
  taxonomic.level = "Phylum",
  top.otu = 3, VariableA = "DiseaseState",
  title = "Rel plot", color = "Set2"
)
print(pn)
}

}
\keyword{utilities}
