% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_cv.R
\name{plot_taxa_cv}
\alias{plot_taxa_cv}
\title{Coefficient of variations}
\usage{
plot_taxa_cv(x, plot.type)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{plot.type}{scatter or hist (histogram)}
}
\value{
A \code{\link{ggplot}} plot object.
}
\description{
Plots CV for OTUs/ASVs.
}
\details{
Check if there are spurious OTUs/ASVs.
}
\examples{
\dontrun{
  # Example data
    library(microbiome)
    library(microbiomeutilities)
    data("zackular2014")
    p0 <- zackular2014
    p <- plot_taxa_cv(p0, plot.type = "hist")
    print(p)
    }

}
\keyword{utilities}
