% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_taxa_heatmap.R
\name{plot_taxa_heatmap}
\alias{plot_taxa_heatmap}
\title{Heatmap using \code{\link{phyloseq-class}} and \code{\link{pheatmap}}}
\usage{
plot_taxa_heatmap(x, subset.top, transformation, VariableA, heatcolors = NULL,
  ...)
}
\arguments{
\item{x}{\code{\link{phyloseq-class}} object.}

\item{subset.top}{either NA or number of Top OTUs to use for plotting.}

\item{transformation}{either 'log10', 'clr','Z', 'compositional', or NA}

\item{VariableA}{main variable of Interest.}

\item{heatcolors}{is the option for colors in \code{\link{pheatmap}}. Default is to use viridis
inferno}

\item{...}{Arguments to be passed \code{\link{pheatmap}}.}
}
\value{
A \code{\link{pheatmap}} plot object.
}
\description{
Plot heatmap using \code{\link{phyloseq-class}} object as input.
}
\examples{

    library(microbiomeutilities)
    library(viridis)
    library(RColorBrewer)
    data("zackular2014")
    ps0 <- zackular2014

    heat.sample <- plot_taxa_heatmap(ps0, subset.top = 20,
                                VariableA = "DiseaseState",
                                heatcolors =brewer.pal(100, "Blues"),
                                transformation = "log10")

}
\author{
Sudarshan A. Shetty (sudarshanshetty9@gmail.com)
}
\keyword{utilities}
